[defaults] Simplify SPC f y d and SPC f y y behavior
SPC f y d now copies returns the true name of default-directory, and
does not look at list-buffers-directory (which is not guaranteed to be
a real directory path and therefore it is inappropriate to pass it to
file-truename).

SPC f y y no longer falls back to using the current directory.